added checkboxes
Dynamic checkboxes for list. Select all works, and unchecking a single checkbox will uncheck the uncheck all. Unchecking the select all box manually will uncheck everything. You know, basic list stuff that everyone knows.

For now, file IDs are dummy data and are in the format of "f<int>" where <int> is just whatever its index in the file list is + 1. This will be replaced with a UID system in the future.